+++ docs/CHANGELOG.md
@@ -10,7 +10,7 @@
 - none
 
 ### improvements
-- none
+- moved features to commons
 
 ### bug fixes
 - none